refactor(design-system): make InlineEditText a preset over useInlineEdit
The focus-and-select, Enter/Escape, trim and skip-unchanged rules existed
twice, in the headless hook and in the design-system component, with nothing
distinguishing them. They now exist once.

The core moved into the design system rather than the component moving out:
the boundaries rule forbids design-system from importing shared, so the shared
copy could never have been the single source. The five callers that render
both modes themselves now import the hook from @/design-system.

Behaviour is unchanged. The empty-input rules were already equivalent: with no
fallback, 'trimmed || fallback || value' commits nothing, which is what
'trimmed && trimmed !== initialValue' did. The component's guard against the
blur a browser fires when Escape unmounts the input moved into the hook, where
it now covers the five hook callers too.

Also drops the LayoutManagerModal re-export shim its two callers still went
through, and records on each of the two APIs when to reach for which.
