+++ src/design-system/InlineEditText/index.ts
@@ -1,2 +1,3 @@
 export { InlineEditText } from './InlineEditText';
 export type { InlineEditTextProps } from './InlineEditText';
+export { useInlineEdit } from './useInlineEdit';
