+++ src/features/baseplate/components/BaseplateLibraryModal/BaseplateLibraryModal.tsx
@@ -17,8 +17,8 @@ import { useMutations } from '@/shared/contexts';
 import { isOk } from '@/core/result';
 import { useTranslation } from '@/i18n';
 import { useMenuKeyboardNav } from '@/shared/hooks/useMenuKeyboardNav';
-import { useInlineEdit, useResponsive } from '@/shared/hooks';
-import { Button, IconButton, Input, XIcon } from '@/design-system';
+import { useResponsive } from '@/shared/hooks';
+import { Button, IconButton, Input, XIcon, useInlineEdit } from '@/design-system';
 import type { BaseplateDesignId } from '@/core/types';
 import type { SavedBaseplateDesign } from '@/features/baseplate/types/library';
 import { listDesigns } from '@/features/baseplate/storage/BaseplateStorage';
