+++ src/features/layout-library/components/LayoutManagerModal/LayoutGridItem/LayoutGridItem.tsx
@@ -1,9 +1,8 @@
 import type { LayoutEntry } from '@/core/types';
 import { LayoutThumbnail } from '@/shell/LayoutThumbnail';
 import { LayoutActions } from '../LayoutActions';
-import { useInlineEdit } from '../useInlineEdit';
 import { useTranslation, useFormatting } from '@/i18n';
-import { Input } from '@/design-system';
+import { Input, useInlineEdit } from '@/design-system';
 
 interface LayoutGridItemProps {
   entry: LayoutEntry;
