--- src/features/layout-library/components/LayoutManagerModal/useInlineEdit.ts
@@ -1,5 +1,0 @@
-/**
- * Re-export from shared hooks for backward compatibility.
- * New code should import directly from @/shared/hooks.
- */
-export { useInlineEdit } from '@/shared/hooks';
