+++ src/shared/hooks/index.ts
@@ -8,7 +8,6 @@ export { usePWAUpdate } from './usePWAUpdate';
 export { usePrefetchChunks } from './usePrefetchChunks';
 export { useFocusTrap } from './useFocusTrap';
 export { useMenuKeyboardNav } from './useMenuKeyboardNav';
-export { useInlineEdit } from './useInlineEdit';
 export { useKeyboard } from './useKeyboard';
 export { usePrefersReducedMotion } from './usePrefersReducedMotion';
